+++ Assets/Resources/Prefabs/Skill/Meteor.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c58c8fbcaf1db3b4a99c02784d3e8f68
+guid: f8b2755f28c67d84fa2dcfdb313db7e3
 PrefabImporter:
   externalObjects: {}
   userData:
