--- Assets/Resources/Prefabs/Skill/Skill.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 593d8e08aa1e87542a6f089904d5ad50
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
