+++ Assets/Scripts/Skill/MeteorSkill.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 37dd6094d8db7274e86667dd5a93a401
+guid: b656e1bc3f33a374c9cb9b79f71c48cb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
