+++ Packages/manifest.json
@@ -13,6 +13,7 @@
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.timeline": "1.7.6",
     "com.unity.ugui": "1.0.0",
+    "com.unity.visualeffectgraph": "14.0.10",
     "com.unity.visualscripting": "1.9.2",
     "com.unity.xr.arcore": "5.1.4",
     "com.unity.xr.arfoundation": "5.1.4",
